Split Single- and MultiplePipelineInput, removed mixins
