+++ metisp/pymetis/src/pymetis/inputs/__init__.py
@@ -19,7 +19,9 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 from .inputset import PipelineInputSet
 
-from .base import PipelineInput, SinglePipelineInput, MultiplePipelineInput
+from .base import PipelineInput
+from .single import SinglePipelineInput
+from .multiple import MultiplePipelineInput
 
 from .common import (RawInput,
                      MasterDarkInput,
