+++ metisp/pymetis/src/pymetis/tests/test_metis_ifu_distortion.py
@@ -36,16 +36,3 @@ def sof():
 class TestRecipe(BaseRecipeTest):
     """ A bunch of extremely simple and stupid test cases... just to see if it does something """
     _recipe = Recipe
-
-    @pytest.mark.skip
-    def test_recipe_can_be_run_directly(self, load_frameset, sof):
-        pass
-
-    @pytest.mark.skip
-    def test_can_be_run_with_pyesorex(self, name, create_pyesorex):
-        pass
-
-    @staticmethod
-    @pytest.mark.skip
-    def test_pyesorex_runs_with_zero_exit_code_and_empty_stderr(name, sof, create_pyesorex):
-        pass
